Split source repository logic into a separate file
Reduces the size of project.py by around 600 lines.
